+++ statistical/calculate_fractions.py
@@ -11,7 +11,7 @@ import rpy2
 import rpy2.robjects as robjects
 import rpy2.robjects.packages as rpackages
 from rpy2.robjects.vectors import DataFrame, StrVector
-from statistical.variance_measures import variance_measures
+from statistical.MOE import variance_measures
 
 survey_package = rpackages.importr("survey")
 base = rpackages.importr("base")
